+++ scripts/check/unit-test.sh
@@ -3,5 +3,5 @@
 set -o errexit
 set -o nounset
 
-go test -race -coverprofile=.test_coverage.txt ./...
+go test -race -coverprofile=.test_coverage.txt $(go list ./... | grep -v /test/integration )
 go tool cover -func=.test_coverage.txt | tail -n1 | awk '{print "Total test coverage: " $3}'